+++ src/auth/auth.service.ts
@@ -6,10 +6,10 @@ import {
 import { JwtService } from "@nestjs/jwt";
 import { InjectModel } from "@nestjs/mongoose";
 import { Model } from "mongoose";
-import { QueryUserDto } from "src/user/dtos";
 
 import { SignupUserDto } from "./dtos";
 import { Session } from "./session.schema";
+import { QueryUserDto } from "../user/dtos";
 import { UserService } from "../user/user.service";
 
 @Injectable()
