+++ src/category/category.module.ts
@@ -1,13 +1,13 @@
 import { Module } from "@nestjs/common";
 import { MongooseModule } from "@nestjs/mongoose";
-import { Session, SessionSchema } from "src/auth/session.schema";
-import { Task, TaskSchema } from "src/task/schemas";
 
 import { CategoryMapperService } from "./category-mapper.service";
 import { CategoryController } from "./category.controller";
 import { Category, CategorySchema } from "./category.schema";
 import { CategoryService } from "./category.service";
 import { AuthModule } from "../auth/auth.module";
+import { Session, SessionSchema } from "../auth/session.schema";
+import { Task, TaskSchema } from "../task/schemas";
 import { User, UserSchema } from "../user/user.schema";
 
 @Module({
