+++ src/category/category.service.ts
@@ -6,7 +6,6 @@ import {
 } from "@nestjs/common";
 import { InjectModel } from "@nestjs/mongoose";
 import { Model, Types } from "mongoose";
-import { Task } from "src/task/schemas";
 
 import { CategoryMapperService } from "./category-mapper.service";
 import { Category } from "./category.schema";
@@ -15,6 +14,7 @@ import {
   CreateCategoryDto,
   QueryCategoryDto,
 } from "./dtos";
+import { Task } from "../task/schemas";
 
 @Injectable()
 export class CategoryService {
