+++ src/image/image.module.ts
@@ -1,10 +1,10 @@
 import { Module } from "@nestjs/common";
 import { MongooseModule } from "@nestjs/mongoose";
-import { AuthModule } from "src/auth/auth.module";
-import { Session, SessionSchema } from "src/auth/session.schema";
 
 import { ImageController } from "./image.controller";
 import { ImageService } from "./image.service";
+import { AuthModule } from "../auth/auth.module";
+import { Session, SessionSchema } from "../auth/session.schema";
 import { User, UserSchema } from "../user/user.schema";
 
 @Module({
