+++ src/task/task.controller.ts
@@ -9,8 +9,6 @@ import {
   Query,
   UseGuards,
 } from "@nestjs/common";
-import { CurrentUser } from "src/auth/decorators";
-import { AuthGuard, BannedUserGuard } from "src/auth/guards";
 
 import { NotificationService } from "./../notification/notification.service";
 import {
@@ -25,6 +23,8 @@ import { SubtaskResponseDto } from "./dtos/response/subtask-response.dto";
 import { SubtaskService } from "./subtask.service";
 import { TaskService } from "./task.service";
 import { UserTasksStats } from "./types";
+import { CurrentUser } from "../auth/decorators";
+import { AuthGuard, BannedUserGuard } from "../auth/guards";
 
 @Controller("task")
 @UseGuards(AuthGuard)
