+++ src/task/task.module.ts
@@ -1,7 +1,5 @@
 import { forwardRef, Module } from "@nestjs/common";
 import { MongooseModule } from "@nestjs/mongoose";
-import { Session, SessionSchema } from "src/auth/session.schema";
-import { CategoryModule } from "src/category/category.module";
 
 import { Subtask, SubtaskSchema, Task, TaskSchema } from "./schemas";
 import { SubtaskMapperService } from "./subtask-mapper.service";
@@ -11,6 +9,8 @@ import { TaskMapperService } from "./task-mapper.service";
 import { TaskController } from "./task.controller";
 import { TaskService } from "./task.service";
 import { AuthModule } from "../auth/auth.module";
+import { Session, SessionSchema } from "../auth/session.schema";
+import { CategoryModule } from "../category/category.module";
 import { Category, CategorySchema } from "../category/category.schema";
 import { NotificationModule } from "../notification/notification.module";
 import { User, UserSchema } from "../user/user.schema";
